split into Generator component
